# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API de Produtos
  description: API para gerenciamento de produtos.
  version: 1.0.0
  contact:
    name: Suporte da API
    url: http://www.example.com/support
    email: suporte@example.com
servers:
  - url: http://localhost:8080/api
    description: Servidor de desenvolvimento

paths:
  /produtos:
    get:
      summary: Listar todos os produtos
      description: Retorna uma lista de todos os produtos cadastrados.
      operationId: listarProdutos
      tags:
        - Produtos
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
        '500':
          description: Erro no servidor ao tentar listar produtos

    post:
      summary: Criar um novo produto
      description: Adiciona um novo produto ao sistema.
      operationId: criarProduto
      tags:
        - Produtos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoInput'
      responses:
        '200':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Dados inválidos fornecidos na solicitação
        '500':
          description: Erro no servidor ao tentar criar o produto

  /produtos/{id}:
    get:
      summary: Buscar produto por ID
      description: Retorna um produto específico baseado no ID.
      operationId: buscarProdutoPorId
      tags:
        - Produtos
      parameters:
        - name: id
          in: path
          required: true
          description: ID do produto a ser buscado
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Produto retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado
        '500':
          description: Erro no servidor ao tentar buscar o produto

components:
  schemas:
    Produto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: "Produto Teste"
        preco:
          type: number
          format: double
          example: 99.99
      required:
        - nome
        - preco

    ProdutoInput:
      type: object
      properties:
        nome:
          type: string